Add `getInvocationSigner` helper function.
